fix: ScaleModes type caster, ConvergenceFlags ordering, bare CUSTOM rejection
- Add ScaleModes enum check in ScaleType type caster so passing
  oc.ScaleModes.AUTO from Python no longer raises TypeError
- Add operator<=> for ConvergenceFlags to formalize severity ordering,
  replacing fragile static_cast<int> comparisons in mesh abort checks
- Reject bare ScaleModes::CUSTOM / "custom" without explicit scale values
  in get_scale_info(), restoring the pre-enum-class error behavior

+++ src/OptimalControl/ODEPhaseBase.cpp
@@ -1567,7 +1567,7 @@ Tycho::PSIOPT::ConvergenceFlags Tycho::ODEPhaseBase::phase_call_impl(JetJobModes
     }
 
     if (this->AdaptiveMesh) {
-        if (static_cast<int>(flag) >= static_cast<int>(this->MeshAbortFlag)) {
+        if (flag >= this->MeshAbortFlag) {
             if (this->PrintMeshInfo) {
                 fmt::print(fmt::fg(fmt::color::red),
                            "Mesh Iteration 0 Failed to Solve: Aborting\n");
@@ -1597,7 +1597,7 @@ Tycho::PSIOPT::ConvergenceFlags Tycho::ODEPhaseBase::phase_call_impl(JetJobModes
                     }
                 }
                 flag = this->psipot_call_impl(nextmode);
-                if (static_cast<int>(flag) >= static_cast<int>(this->MeshAbortFlag)) {
+                if (flag >= this->MeshAbortFlag) {
                     if (this->PrintMeshInfo) {
                         fmt::print(fmt::fg(fmt::color::red),
                                    "Mesh Iteration {0:} Failed to Solve: Aborting\n", i + 1);
